feat(configure): simplify configure
1. remove custom level
2. add global configure checker function
3. appender config merge to category configure
