+++ src/configure/index.ts
@@ -1,0 +1,1 @@
+export { configure } from './configure';
