+++ src/configure/setup.ts
@@ -1,0 +1,5 @@
+import { Configuration } from './configure';
+
+export const setup = (configuration: Configuration): void => {
+  return;
+};
